docker-enrichment/requirements2.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
blis==0.7.9
boto3==1.26.94
botocore==1.29.94
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cryptography==39.0.2
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
fastcoref
filelock==3.10.0
finbert-embedding==0.1.5
flatbuffers==23.3.3
frozenlist==1.3.3
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.13.2
idna==3.4
isodate==0.6.1
jax==0.4.6
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keras==2.11.0
langcodes==3.3.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
mpmath==1.3.0
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
murmurhash==1.0.9
networkx==3.0
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pathy==0.10.1
portalocker
preshed
protobuf
pyasn1
pyasn1-modules==0.2.8
pycparser==2.21
pydantic
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch-pretrained-bert==0.6.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
simplejson==3.18.4
six==1.16.0
smart-open==6.3.0
spacy
spacy-legacy
spacy-loggers
srsly==2.4.6
sympy==1.11.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
thinc==8.1.9
threadpoolctl==3.1.0
tokenizers==0.13.2
torch
tqdm
transformers
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==0.10.1
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.8.2
